Moved CloudCityUtil to new 'util' package as well
